+++ headers/sequential/PrimePathSequential.h
@@ -6,10 +6,10 @@ class PrimePathSequential: public IPathBuilder {
 public:
     void displayPath(const std::string &start, const std::string &end, Graph *graph);
 private:
-    bool isPrime(int n);
+    bool isPrime(size_t n);
     struct State {
         std::string node;
-        int weight;
+        size_t weight;
         std::vector<std::string> path;
     };
 };
