+++ headers/sequential/ShortestPrimePathSequential.h
@@ -7,10 +7,10 @@ public:
     void displayPath(const std::string &start, const std::string &end, Graph *graph);
 
 private:
-    bool isPrime(int n);
+    bool isPrime(size_t n);
     struct State {
         std::string node;
-        int weight;
+        size_t weight;
         std::vector<std::string> path;
     };
     struct StateComparator;
