+++ headers/threaded/PrimePathThreaded.h
@@ -9,7 +9,7 @@ private:
     static bool isPrime(int n);
     struct State {
         std::string node;
-        int weight;
+        size_t weight;
         std::vector<std::string> path;
     };
 };
