+++ headers/threaded/ShortestPrimePathThreaded.h
@@ -9,11 +9,11 @@ public:
     void displayPath(const std::string &start, const std::string &end, Graph* graph) override;
     struct State {
         std::string node;
-        int weight;
+        size_t weight;
         std::vector<std::string> path;
     };
 private:
-    static bool isPrime(int n);
+    static bool isPrime(size_t n);
 
 };
 #endif //SHORTESTPRIMEPATHTHREADED_H
